Fix z-fighting in the soul cairn
FXSmokeWispsLg2x1Adjust was used at the top of one of the tallest soul
cairn towers. Swapping it was causing z-fighting with that tower. I
tested two other locations that FXSmokeWispsLg2x1Adjust was used (one
exterior Tamriel, one in an interior dungeon) and couldn't reproduce the
z-fighting in those locations. This means it's likely fine to continue
to swap that mesh in general, just not in the soul cairn.

Only swap FXSmokeWispsLg2x1Adjust|EmptyFX01 with non soul cairn
locations, which fixes the z-fighting.

Thanks to Skyrimer3232 on the Nexus for the report and finding out what
swap was causing the bug.