Use lp_overhaul.esp to provide InverseSquareRootLight
Instead of using Bottle's inverse_square_root_lights.esp, create a new
ESL flagged ESPFE which creates one new light called
InverseSquareRootLight. It has the flags "Dynamic", and "Unknown Flag
14" which causes CS to treat it as an inverse square root attenuating
light.

Change all lights in all LP configs to use it.

This is being done because it's easier to audit which lights are flagged
as InverseSquareRoot when they're all the same light. Since Light Placer
can define all properties of a light, the only other reason to have more
than one light is to provide something like a shadow hemisphere or
spotlight, but those aren't needed for LPO at this time. Perhaps they
may be added in the future.